Add the initial pyspace
This is an experiment to reimplement the KnownSpace kernel from
fading memory in Python.

The basic idea is that there is a thing called a Pool in which exist
Entitities and Events. Observing the Pool are Simpletons, things
which "hear" Events and do stuff, simple stuff, in response.

Thus far this implementation is not very much like KnownSpace. Instead
it is just a tool to play with Threads. One can 'make run' to watch it
do something very limited:

* create a PriorityQueue
* start a KernelSimpleton and fire a KernelEvent telling it init
* start up N DebugSimpleton's
* read a file and send Events for each line, the DebugSimpleton's will
  pick this up

The most important difference between this and KnownSpace is that the
event is consumed off the queue after its first use rather than being
heard by everything that cares about it. That's because addressing that
problem wasn't a first concern. It'll come.

So too will Entities, which allow stuff to be persisted. Will probably
just do that as files in a dir, initially.

What you end up with is a lot of Event and Simpleton subclasses
spinning around in the pool in a tumbling cascade of Events. At first
glance the code seems annoyingly verbose for little achievement, but
what it models is free agency in/on a broad pool, but without actually
having to worry about that broad pool.